[REFACT] ImageController 추가 및 패키지 변경
